Allow title and button-based appender to inherit styles.
File under "should've probably done this sooner", this PR allows the title field and the initial appender to inherit the theme styles defined by an editor style.

In the past this wasn't possible due to how things were structured. But as it is, those styles are inherited by what is applied to the `editor-styles-wrapper`, which means this works both for themes that don't register and editor style — it looks like before — but now it also works for themes that do.